Compute the likelihood.

This function computes the likelihood associated to a specific set of parameters. The specification of parameters determines the ecological processes involved.

getLikelihood(pars, obs, zero_infl = FALSE, neigh = NULL,
  disp = NULL, favo = NULL, SDBH = NULL, kernel = "kern_lognormal",
  pstr = NA, ppz = NA, pscal = NA, pshap = NA, pfav = NA,
  pneigh = NA, quiet = TRUE, record = NULL)

Arguments

pars
vector of parameters values, the size must

obs
vector of observations of recruitement for each plots of the stand.

zero_infl
logical. If TRUE then a zero-inflated poisson distribution is used, otherwose a poisson distribution is used.

neigh
a logical. If NULL (the default setting) the is TRUE then neighborhood

disp
a logical. If NULL (the default setting) the is TRUE then neighborhood

favo
a logical. If NULL (the default setting) the is TRUE then neighborhood

SDBH
a numeric vector of standarized DBH (Diameter at breast heights).

kernel
a character string indicating which kernel should be used (either kern_lognormal or kern_exponential_power). See kernels for further details.

pstr
position of paraneter in thevector of parameter values.

ppz
the position of pz parameter within the vector of parameter values.

pscal
An integer gving the position of the scale parameter (of the dispersal kernel) witin the vector of parameter values.

pshap
An integer gving the position of the shape parameter (of the dispersal kernel) witin the vector of parameter values.

pfav
A vector of integer indicating the position of the five favorability values within the vector of parameter values.

pneigh
position of the parameter associated to the neighborhood effect in the vector of parameter values.

quiet
logial. If FALSE then parameters and likelyhood estimations are printed.

record
a connection, or a character string naming the file to print to. If NULL, the default values, no record is done.

Value

The log-likelihood values associated to a given set of parameters.

Details

The parameter pz have a value even if it is not included in the analysis (meaning a poisson distrubution is used).