Recruitement analayis .

This function computes the likelihood associated to a specific set of parameters. The specification of parameters determines the ecological processes involved.

recuitmentAnalysis(site, year, tree, age, mat_par, path = "./",
  zero_infl = FALSE, disp = FALSE, favo = FALSE, neigh = FALSE,
  clip = NULL, kernel = "kern_lognormal", mxt = 100, quiet = TRUE,
  record = NULL)

launchIt(iter, simu = NULL, simu_file = "dataReady/simu_all.Rds",
  mxt = 100, record = "test.txt", quiet = TRUE,
  path = "dataReady/")

Arguments

site
site name.

year
year (2015 or 2016) name.

tree
abbreviation of the tree species names of the species.

age
age of the species (1 or 2).

mat_par
a matrix of parameters (see getParameters).

path
a character string giving the path to access the data.

zero_infl
logical. If TRUE then a zero-inflated poisson distribution is used, otherwose a poisson distribution is used.

disp
a logical. If TRUE, then dispersal parameters are used.

favo
a logical. If TRUE, then favorability parameters are used.

neigh
a logical. If TRUE, then neighborhood effect parameter is used.

clip
a numeric vector of standarized DBH (Diameter at breast heights).

kernel
a character string indicating which kernel should be used (either kern_lognormal or kern_exponential_power). See kernels for further details.

mxt
a numeric. Maximum running time in seconds (see GenSA).

quiet
logial. If TRUE details about the state of simulations are printed.

record
a connection, or a character string naming the file to print to. If NULL, the default values, no record is done.

iter
a interger indentifying the iteration.

simu
a data frame describing the simulation.

simu_file
a file that includes a description of simulation.

Value

The log-likelihood values associated to a given set of parameters.

Details

The parameter pz have a value even if it is not included in the analysis (meaning a poisson distrubution is used).

Functions

launchIt: Launch the analysis for a given row of the simulation data frame or file.